pyproject.toml
[tool.poetry]
name = "drf-auth"
version = "0.1.0"
description = ""
authors = ["anas-abusaif <anasabusief@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.9"
djangorestframework = "^3.12.4"
djangorestframework-simplejwt = "^5.0.0"
httpie = "^2.6.0"
django-cors-headers = "^3.10.0"
gunicorn = "^20.1.0"
whitenoise = "^5.3.0"
psycopg2-binary = "^2.9.2"

[tool.poetry.dev-dependencies]
flake8 = "^4.0.1"
black = "^21.11b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
